% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_env.R
\name{norm_env}
\alias{norm_env}
\title{Normal Lake Stage Envelope Score}
\usage{
norm_env(stg.data, allinfo = FALSE)
}
\arguments{
\item{stg.data}{see details}

\item{allinfo}{TRUE/FALSE if TRUE, the function will return values used in computing score}
}
\value{
Returns a \code{data.frame} of original data and normal stage elevation score.
}
\description{
Calculates a normal lake stage envelope score
}
\details{
The input \code{stg.data} is a \code{data.frame} with columns:
\itemize{
\item{\code{Date} (as a \code{POSIXct} or \code{Date} variable)}
\item{\code{Data.Value} as stage elevation data in feet (NGVD29)}
}
}
\examples{
# Example dataset (not real data)
dates=seq(as.Date("2016-01-01"),as.Date("2016-02-02"),"1 days")
dat=data.frame(Date=dates,Data.Value=runif(33,12,18))

score=norm_env(dat)
# END
}
