% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_stage_score.R
\docType{data}
\name{rec.stage.score}
\alias{rec.stage.score}
\title{Recovery Lake Stage Envelope thresholds}
\format{
An object of class \code{data.frame} with 366 rows and 11 columns.
}
\source{
RECOVER, 2020, Lake Okeechobee Performance Measure - Lake State
}
\usage{
data(rec.stage.score)
}
\description{
Thresholds for the recovery lake stage envelope to approximate corresponding scores for lake stages

Values are represented as Feet NGVD29
}
\keyword{datasets}
